OM-53397: Update reservations status when the depended template is deleted (rb36439)
    Problem:
    ==========
    Currently if templates are deleted (both user and system), we don't handle the reservations that use these templates (it's a TODO).
    So reservations will fail due to missing template.

    Proposed solution:
    =================
    After discussed in the team, we decided to:
    User templates - notify the user there are reservations using the template, and stop the operation. In this change, UI will only show operation failed. Additional UX work will be done on 54227.
    System templates - Set reservations to invalid when system templates are removed due to targets are deleted.

    Testing Done
    Added new unit tests for the changes.
    All existing unit tests passed.

+++ com.vmturbo.plan.orchestrator/src/main/java/com/vmturbo/plan/orchestrator/templates/TemplatesRpcService.java
@@ -12,6 +12,8 @@ import javax.annotation.Nonnull;
 
 import com.google.common.collect.Iterators;
 
+import com.vmturbo.common.protobuf.plan.ReservationDTO;
+import com.vmturbo.plan.orchestrator.reservation.ReservationDao;
 import io.grpc.Status;
 import io.grpc.stub.StreamObserver;
 
@@ -53,11 +55,15 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
 
     private final int templateChunkSize;
 
+    private final ReservationDao reservationDao;
+
     TemplatesRpcService(@Nonnull final TemplatesDao templatesDao,
                         @Nonnull final DeploymentProfileDaoImpl deploymentProfileDao,
+                        @Nonnull final ReservationDao reservationDao,
                         final int templateChunkSize) {
         this.templatesDao = Objects.requireNonNull(templatesDao);
         this.deploymentProfileDao = Objects.requireNonNull(deploymentProfileDao);
+        this.reservationDao = Objects.requireNonNull(reservationDao);
         this.templateChunkSize = templateChunkSize;
     }
 
@@ -70,6 +76,7 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
                 .withDescription("Delete templates by target ID must have an target ID").asException());
             return;
         }
+
         try {
             final List<Template> deletedTemplates = templatesDao.deleteTemplateByTargetId(request.getTargetId());
             deletedTemplates.stream().forEach(responseObserver::onNext);
@@ -210,6 +217,13 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
                 .withDescription("Delete template must have an template ID").asException());
             return;
         }
+
+        if (templateUsedByReservation(request.getTemplateId())) {
+            responseObserver.onError(Status.FAILED_PRECONDITION
+                    .withDescription("Delete the reservations used by this template first.").asException());
+            return;
+        }
+
         try {
             final Template template = templatesDao.deleteTemplateById(request.getTemplateId());
             responseObserver.onNext(template);
@@ -227,6 +241,17 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
                 .withDescription(e.getLocalizedMessage())
                 .asException());
         }
+    }
+
+    private boolean templateUsedByReservation(long templateId) {
+        final Set<ReservationDTO.Reservation> reservationsByTemplates =
+                reservationDao.getReservationsByTemplates(Collections.singleton(templateId));
+        if (!reservationsByTemplates.isEmpty()) {
+            logger.error("This template {} is being used by reservations {}.", templateId,
+                    reservationsByTemplates.stream().map(ReservationDTO.Reservation::getName).collect(Collectors.toSet()));
+            return true;
+        }
+        return false;
     }
 
     @Override
